Fix Linters errors
