+++ config/database.yml
@@ -62,8 +62,8 @@ test:
   <<: *default
   database: Reciepe_App_test
   port: 5433
-  username: postgres
-  password: 12345678
+  username: teejay
+  password: rojutet11
 
 # As with config/credentials.yml, you never want to store sensitive information,
 # like your database password, in your source code. If your source code is
