+++ config/routes.rb
@@ -1,5 +1,5 @@
 Rails.application.routes.draw do
-  root 'users#index'  
+  root 'foods#index'  
   resources :users, only: [:index, :show] do
     resources :foods, only: [:index, :show, :new, :create, :destroy] do
       resources :recipes, only: [:index, :create, :destroy]
